Condensing the application into one file, changing auth token, github token
Everything now should be run securely in production.

Authorization tokens must be retrieved by calling '/token'. That will return a token (which really isn't used, but could be) and a header called base_token which is used to generate the following tokens.

GitHub authorization tokens should be done a little better now as well.